+++ src/module/Members/member.service.ts
@@ -1,5 +1,7 @@
 import { Prisma } from "@prisma/client";
 import { prisma } from "../../config/prisma";
+import AppError from "../../utils/AppError";
+import httpStatus from "http-status";
 
 const createMemberDb = async (memberDetail: any) => {
   const member: Prisma.MemberCreateInput = await prisma.member.create({
@@ -13,22 +15,29 @@ const getMembersDb = async () => {
   return members;
 };
 const getSingleMemberDb = async (id: string) => {
-  const member: Prisma.MemberCreateInput =
-    await prisma.member.findUniqueOrThrow({
-      where: {
-        memberId: id,
-      },
-    });
+  const member = await prisma.member.findUnique({
+    where: {
+      memberId: id,
+    },
+  });
+
+  if (!member) {
+    throw new AppError(httpStatus.BAD_REQUEST, "No Member Matches this ID!");
+  }
 
   return member;
 };
 
 const updateMemberDb = async (id: string, updatedData: any) => {
-  const isExist = await prisma.member.findUniqueOrThrow({
+  const isExist = await prisma.member.findUnique({
     where: {
       memberId: id,
     },
   });
+
+  if (!isExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "No Member Matches this ID!");
+  }
 
   const updated = await prisma.member.update({
     where: {
@@ -51,6 +60,9 @@ const deleteMemberDb = async (id: string) => {
     },
   });
 
+  if (!isExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "No Member Matches this ID!");
+  }
   const deletedData = await prisma.member.delete({
     where: {
       memberId: id,
